+++ NEWS.md
@@ -1,4 +1,6 @@
 # Version 1.0.9000.9030
+* Renamed generic function and method `biomLevel` to `biomarker`.
+  Changed its arguments.
 * Renamed several slots in `NextBestNCRM` and `NextBestDualEndpoint` classes.
 * Changed the `scale` character slot into `target_relative` flag slot in
   `NextBestDualEndpoint` class.